[ADD] changed UserHasHouseshareEntity isOwner to int (easier like this) -> finished HouseShare.java path, user can now create new houseshare, and list his houseshare when he is logged in.

+++ src/main/java/com/example/sharelockapi/controllers/UserHasHouseShareManager.java
@@ -25,7 +25,7 @@ public class UserHasHouseShareManager {
 
 
 
-    public static boolean createUserHasHouseShare(UserEntity user, HouseshareEntity houseshare, int points, Byte isOwner) {
+    public static boolean createUserHasHouseShare(UserEntity user, HouseshareEntity houseshare, int points, int isOwner) {
         UserHasHouseshareEntity u = daoUserHasHouseShare.find(user.getId());
         if (u == null) {
             DAOUserHasHouseShare.create(
